added : 實作掉落方塊以等級分別有不同的掉落速度，但未實際加上升級條件。

+++ src/tetris/view/InfoBar.java
@@ -7,7 +7,7 @@ package tetris.view;
  */
 public class InfoBar {
 	private long mScore;// 遊戲分數
-	private short mLevel;// 遊戲等級
+	private int mLevel;// 遊戲等級
 	private int mCleanedCount;// 已消除的行數
 
 	public InfoBar() {
@@ -32,15 +32,15 @@ public class InfoBar {
 		mScore += score;
 	}
 
-	public short getLevel() {
+	public int getLevel() {
 		return mLevel;
 	}
 
-	public void setLevel(short level) {
+	public void setLevel(int level) {
 		mLevel = level;
 	}
 
-	public void addLevel(short level) {
+	public void addLevel(int level) {
 		mLevel += level;
 	}
